+++ R/Connect.R
@@ -31,7 +31,8 @@ checkIfDbmsIsSupported <- function(dbms) {
     "sqlite extended",
     "spark",
     "snowflake",
-    "synapse"
+    "synapse",
+    "duckdb"
   )
   if (!dbms %in% supportedDbmss) {
     abort(sprintf(
@@ -279,6 +280,8 @@ connect <- function(connectionDetails = NULL,
     
     if (connectionDetails$dbms %in% c("sqlite", "sqlite extended")) {
       connectSqlite(connectionDetails)
+    } else if (connectionDetails$dbms == "duckdb") {
+      connectDuckdb(connectionDetails)
     } else if (connectionDetails$dbms == "spark" && is.null(connectionDetails$connectionString())) {
       connectSparkUsingOdbc(connectionDetails)
     } else {
@@ -636,14 +639,22 @@ connectSpark <- function(connectionDetails) {
   inform("Connecting using Spark JDBC driver")
   jarPath <- findPathToJar("^SparkJDBC42\\.jar$", connectionDetails$pathToDriver)
   driver <- getJbcDriverSingleton("com.simba.spark.jdbc.Driver", jarPath)
-  if (is.null(connectionDetails$connectionString()) || connectionDetails$connectionString() == "") {
+  connectionString <- connectionDetails$connectionString()
+  if (is.null(connectionString) || connectionString == "") {
     abort("Error: Connection string required for connecting to Spark.")
   }
+  if (!grepl("UseNativeQuery", connectionString)) {
+    if (!endsWith(connectionString, ";")) {
+      connectionString <- paste0(connectionString, ";")
+    }
+    connectionString <- paste0(connectionString, "UseNativeQuery=1")
+  }
   if (is.null(connectionDetails$user())) {
-    connection <- connectUsingJdbcDriver(driver, connectionDetails$connectionString(), dbms = connectionDetails$dbms)
+    connection <- connectUsingJdbcDriver(driver, connectionString, dbms = connectionDetails$dbms)
   } else {
-    connection <- connectUsingJdbcDriver(driver,
-      connectionDetails$connectionString(),
+    connection <- connectUsingJdbcDriver(
+      jdbcDriver = driver,
+      url = connectionString,
       user = connectionDetails$user(),
       password = connectionDetails$password(),
       dbms = connectionDetails$dbms
@@ -683,13 +694,15 @@ connectSnowflake <- function(connectionDetails) {
     abort("Error: Connection string required for connecting to Snowflake.")
   }
   if (is.null(connectionDetails$user())) {
-    connection <- connectUsingJdbcDriver(driver, connectionDetails$connectionString(), dbms = connectionDetails$dbms)
+    connection <- connectUsingJdbcDriver(driver, connectionDetails$connectionString(), dbms = connectionDetails$dbms,
+                    "CLIENT_TIMESTAMP_TYPE_MAPPING"="TIMESTAMP_NTZ")
   } else {
     connection <- connectUsingJdbcDriver(driver,
       connectionDetails$connectionString(),
       user = connectionDetails$user(),
       password = connectionDetails$password(),
-      dbms = connectionDetails$dbms
+      dbms = connectionDetails$dbms,
+      "CLIENT_TIMESTAMP_TYPE_MAPPING"="TIMESTAMP_NTZ"
     )
   }
   return(connection)
@@ -769,6 +782,20 @@ connectUsingDbi <- function(dbiConnectionDetails) {
   return(connection)
 }
 
+connectDuckdb <- function(connectionDetails) {
+  inform("Connecting using DuckDB driver")
+  ensure_installed("duckdb")
+  connection <- connectUsingDbi(
+    createDbiConnectionDetails(
+      dbms = connectionDetails$dbms,
+      drv = duckdb::duckdb(),
+      dbdir = connectionDetails$server(),
+      bigint = "integer64"
+    )
+  )
+  return(connection)
+}
+
 generateRandomString <- function(length = 20) {
   return(paste(sample(c(letters, 0:9), length, TRUE), collapse = ""))
 }
@@ -810,7 +837,11 @@ disconnect.default <- function(connection) {
 
 #' @export
 disconnect.DatabaseConnectorDbiConnection <- function(connection) {
-  DBI::dbDisconnect(connection@dbiConnection)
+  if (connection@dbms == "duckdb") {
+    DBI::dbDisconnect(connection@dbiConnection, shutdown = TRUE)
+  } else {
+    DBI::dbDisconnect(connection@dbiConnection)
+  }
   unregisterWithRStudio(connection)
   invisible(TRUE)
 }