+++ R/HelperFunctions.R
@@ -139,3 +139,7 @@ extractQueryTimesFromThread <- function(thread, logLines) {
   return(result) 
 }
 
+supportsAutoCommit <- function(dbms) {
+  return(dbms != "spark")
+}
+