+++ R/Sql.R
@@ -207,7 +207,8 @@ lowLevelQuerySql.default <- function(connection,
     "org.ohdsi.databaseConnector.BatchedQuery",
     connection@jConnection,
     query,
-    dbms(connection)
+    dbms(connection),
+    supportsAutoCommit(dbms(connection))
   )
   
   on.exit(rJava::.jcall(batchedQuery, "V", "clear"))
@@ -225,7 +226,6 @@ lowLevelQuerySql.default <- function(connection,
                                  integer64AsNumeric = integer64AsNumeric,
                                  integerAsNumeric = integerAsNumeric
     )
-    
     columns <- rbind(columns, batch)
   }
   delta <- Sys.time() - startTime
@@ -428,9 +428,10 @@ executeSql <- function(connection,
   }
   
   startTime <- Sys.time()
+  dbms <- dbms(connection)
   
   if (inherits(connection, "DatabaseConnectorJdbcConnection") &&
-      dbms(connection) == "redshift" &&
+      dbms == "redshift" &&
       rJava::.jcall(connection@jConnection, "Z", "getAutoCommit")) {
     # Turn off autocommit for RedShift to avoid this issue:
     # https://github.com/OHDSI/DatabaseConnector/issues/90
@@ -468,7 +469,7 @@ executeSql <- function(connection,
           logTrace(paste("Statements took", delta, attr(delta, "units")))
         },
         error = function(err) {
-          .createErrorReport(dbms(connection), err$message, paste(batchSql, collapse = "\n\n"), errorReportFile)
+          .createErrorReport(dbms, err$message, paste(batchSql, collapse = "\n\n"), errorReportFile)
         },
         finally = {
           rJava::.jcall(statement, "V", "close")
@@ -497,7 +498,7 @@ executeSql <- function(connection,
           }
         },
         error = function(err) {
-          .createErrorReport(dbms(connection), err$message, sqlStatement, errorReportFile)
+          .createErrorReport(dbms, err$message, sqlStatement, errorReportFile)
         }
       )
       if (progressBar) {
@@ -509,7 +510,8 @@ executeSql <- function(connection,
     }
   }
   
-  if (inherits(connection, "DatabaseConnectorJdbcConnection") && !rJava::.jcall(connection@jConnection, "Z", "getAutoCommit")) {
+  if (inherits(connection, "DatabaseConnectorJdbcConnection") && 
+      (!supportsAutoCommit(dbms) || !rJava::.jcall(connection@jConnection, "Z", "getAutoCommit"))) {
     rJava::.jcall(connection@jConnection, "V", "commit")
   }
   
@@ -540,8 +542,8 @@ convertFields <- function(dbms, result) {
       }
     }
   }
-  if (dbms %in% c("bigquery")) {
-    # BigQuery doesn't have INT fields, only INT64. For more consistent behavior with other
+  if (dbms %in% c("bigquery", "snowflake")) {
+    # BigQuery and Snowflake don't have INT fields, only INT64. For more consistent behavior with other
     # platforms, if it fits in an integer, convert it to an integer:
     if (ncol(result) > 0) {
       for (i in 1:ncol(result)) {
@@ -610,7 +612,6 @@ querySql <- function(connection,
   if (inherits(connection, "DatabaseConnectorJdbcConnection") && rJava::is.jnull(connection@jConnection)) {
     abort("Connection is closed")
   }
-
   # Calling splitSql, because this will also strip trailing semicolons (which cause Oracle to crash).
   sqlStatements <- SqlRender::splitSql(sql)
   if (length(sqlStatements) > 1) {