+++ NEWS.md
@@ -1,3 +1,25 @@
+DatabaseConnector 6.1.0
+=======================
+
+Changes:
+
+1. Adding support for DuckDb
+
+Bugfixes:
+
+1. Fixed `capacity < 0` error message when using a large Java heap space.
+
+2. Fixed 'optional feature not supported' error when connecting to DataBricks using JDBC.
+
+3. Fixed `insertTable()` on Snowflake when data includes `POSIXct` type.
+
+4. Fixed 'out of Java heap space' when fetching data with (large) strings. This is achieved by checking the available Java heap space at every 10,000 rows, stopping the batch when less than half is still available. Additionally, all strings from the previous batch are de-referenced before starting a new batch.
+
+5. Fixing DATETIME shifts on Snowflake.
+
+6. Fixing unit tests for upcoming Andromeda version (using arrow instead of SQLite).
+
+
 DatabaseConnector 6.0.0
 =======================